[refactor] _core.host_representation.repository_location -> code_location
