+++ python_modules/dagster-graphql/dagster_graphql/implementation/fetch_assets.py
@@ -44,9 +44,9 @@ from dagster._core.definitions.time_window_partitions import (
     fetch_flattened_time_window_ranges,
 )
 from dagster._core.events import ASSET_EVENTS
+from dagster._core.host_representation.code_location import CodeLocation
 from dagster._core.host_representation.external import ExternalRepository
 from dagster._core.host_representation.external_data import ExternalAssetNode
-from dagster._core.host_representation.repository_location import CodeLocation
 from dagster._core.instance import DynamicPartitionsStore
 from dagster._core.storage.partition_status_cache import (
     can_cache_partition_type,
