+++ python_modules/dagster-graphql/dagster_graphql/implementation/fetch_env_vars.py
@@ -1,6 +1,6 @@
 import dagster._check as check
 from dagster._core.definitions.selector import RepositorySelector
-from dagster._core.host_representation.repository_location import CodeLocation
+from dagster._core.host_representation.code_location import CodeLocation
 from graphene import ResolveInfo
 
 from dagster_graphql.schema.env_vars import (
