+++ python_modules/dagster-graphql/dagster_graphql_tests/graphql/test_context.py
@@ -5,7 +5,7 @@ from unittest import mock
 
 import pytest
 from dagster import op, repository
-from dagster._core.host_representation.repository_location import GrpcServerCodeLocation
+from dagster._core.host_representation.code_location import GrpcServerCodeLocation
 from dagster._core.test_utils import instance_for_test
 from dagster._legacy import pipeline
 from dagster_graphql.test.utils import define_out_of_process_workspace, main_repo_location_name
@@ -39,7 +39,7 @@ def test_can_reload_on_external_repository_error():
                 # note it where the function is *used* that needs to mocked, not
                 # where it is defined.
                 # see https://docs.python.org/3/library/unittest.mock.html#where-to-patch
-                "dagster._core.host_representation.repository_location.sync_get_streaming_external_repositories_data_grpc"
+                "dagster._core.host_representation.code_location.sync_get_streaming_external_repositories_data_grpc"
             ) as external_repository_mock:
                 external_repository_mock.side_effect = Exception("get_external_repo_failure")
