+++ python_modules/dagster-graphql/dagster_graphql_tests/graphql/test_reload_repository_location.py
@@ -284,13 +284,13 @@ class TestReloadRepositoriesOutOfProcess(OutOfProcessTestSuite):
             # note it where the function is *used* that needs to mocked, not
             # where it is defined.
             # see https://docs.python.org/3/library/unittest.mock.html#where-to-patch
-            "dagster._core.host_representation.repository_location.sync_list_repositories_grpc"
+            "dagster._core.host_representation.code_location.sync_list_repositories_grpc"
         ) as cli_command_mock:
             with mock.patch(
                 # note it where the function is *used* that needs to mocked, not
                 # where it is defined.
                 # see https://docs.python.org/3/library/unittest.mock.html#where-to-patch
-                "dagster._core.host_representation.repository_location.sync_get_streaming_external_repositories_data_grpc"
+                "dagster._core.host_representation.code_location.sync_get_streaming_external_repositories_data_grpc"
             ) as external_repository_mock:
 
                 @repository
@@ -351,7 +351,7 @@ class TestReloadRepositoriesOutOfProcess(OutOfProcessTestSuite):
             # note it where the function is *used* that needs to mocked, not
             # where it is defined.
             # see https://docs.python.org/3/library/unittest.mock.html#where-to-patch
-            "dagster._core.host_representation.repository_location.sync_list_repositories_grpc"
+            "dagster._core.host_representation.code_location.sync_list_repositories_grpc"
         ) as cli_command_mock:
             cli_command_mock.side_effect = Exception("Mocked repository load failure")
