+++ python_modules/dagster/dagster/_core/host_representation/__init__.py
@@ -6,6 +6,11 @@ It also contains classes that represent historical representations
 that have been persisted. e.g. HistoricalPipeline
 """
 
+from .code_location import (
+    CodeLocation as CodeLocation,
+    GrpcServerCodeLocation as GrpcServerCodeLocation,
+    InProcessCodeLocation as InProcessCodeLocation,
+)
 from .external import (
     ExternalExecutionPlan as ExternalExecutionPlan,
     ExternalPartitionSet as ExternalPartitionSet,
@@ -52,9 +57,4 @@ from .origin import (
     ManagedGrpcPythonEnvCodeLocationOrigin as ManagedGrpcPythonEnvCodeLocationOrigin,
 )
 from .pipeline_index import PipelineIndex as PipelineIndex
-from .repository_location import (
-    CodeLocation as CodeLocation,
-    GrpcServerCodeLocation as GrpcServerCodeLocation,
-    InProcessCodeLocation as InProcessCodeLocation,
-)
 from .represented import RepresentedPipeline as RepresentedPipeline
