+++ python_modules/dagster/dagster/_core/host_representation/handle.py
@@ -9,7 +9,7 @@ from dagster._core.host_representation.origin import (
 from dagster._core.origin import RepositoryPythonOrigin
 
 if TYPE_CHECKING:
-    from dagster._core.host_representation.repository_location import CodeLocation
+    from dagster._core.host_representation.code_location import CodeLocation
 
 
 class RepositoryHandle(
@@ -24,7 +24,7 @@ class RepositoryHandle(
     )
 ):
     def __new__(cls, repository_name: str, code_location: "CodeLocation"):
-        from dagster._core.host_representation.repository_location import CodeLocation
+        from dagster._core.host_representation.code_location import CodeLocation
 
         check.inst_param(code_location, "code_location", CodeLocation)
         return super(RepositoryHandle, cls).__new__(
