+++ python_modules/dagster/dagster_tests/daemon_tests/test_queued_run_coordinator_daemon.py
@@ -4,9 +4,9 @@ from typing import Iterator
 
 import pytest
 from dagster._core.events import DagsterEvent, DagsterEventType
+from dagster._core.host_representation.code_location import GrpcServerCodeLocation
 from dagster._core.host_representation.handle import JobHandle
 from dagster._core.host_representation.origin import ManagedGrpcPythonEnvCodeLocationOrigin
-from dagster._core.host_representation.repository_location import GrpcServerCodeLocation
 from dagster._core.storage.pipeline_run import IN_PROGRESS_RUN_STATUSES, DagsterRunStatus
 from dagster._core.storage.tags import PRIORITY_TAG
 from dagster._core.test_utils import (
