+++ python_modules/dagster/dagster_tests/general_tests/grpc_tests/test_grpc_server_registry.py
@@ -5,12 +5,12 @@ import time
 import pytest
 from dagster import file_relative_path, repository
 from dagster._core.errors import DagsterUserCodeProcessError
+from dagster._core.host_representation.code_location import GrpcServerCodeLocation
 from dagster._core.host_representation.grpc_server_registry import GrpcServerRegistry
 from dagster._core.host_representation.origin import (
     ManagedGrpcPythonEnvCodeLocationOrigin,
     RegisteredCodeLocationOrigin,
 )
-from dagster._core.host_representation.repository_location import GrpcServerCodeLocation
 from dagster._core.test_utils import instance_for_test
 from dagster._core.types.loadable_target_origin import LoadableTargetOrigin
 from dagster._legacy import pipeline
